Move fixtures to let consumers declare them as `pytest_plugins`
Pytest docs recommend against importing them. Instead, they should be
declared as plugins via something like

```
pytest_plugins = "datalad_next.tests.fixtures"
```

https://docs.pytest.org/en/latest/how-to/fixtures.html#using-fixtures-from-other-projects
